Added tm_thread_block_schedule_work
This call is designed to provide an atomic way to set up a thread for blocking, and then have some work executed once it finally schedules away. This is a common pattern in, for example, the block layer, where we would enqueue a request to the elevator while being in a partially blocked state in order to avoid a race condition.