Add THREAD_WAKEUP flag and tm_thread_poke
This fixes a possible race condition where a worker thread might
choose to sleep due to lack of work, right as another thread tries
to assign it something to do. Specifically, the thread may set its
state to THREADSTATE_INTERRUPTIBLE just as another thread attempts
to wake it up, making the assumption that it might be asleep.

The tm_thread_poke call is introduced as the "safe" way to wake up
a thread which might choose to sleep. It sets an additional flag,
THREAD_WAKEUP, which the scheduler looks at when scheduling away
from a thread. If that flag is set, and the thread has state set
as THREADSTATE_INTERRUPTIBLE, then the state is reset to RUNNING,
and the THREAD_WAKEUP flag is cleared. The ensures that a thread
that is trying to sleep while another thread wakes it up will
always return from tm_thread_set_state with the state set to
RUNNING.