Make valloc's next fit location update during region release.
Since the majority of things that use valloc allocate one 'page'
during an allocation, next-fit can be greatly improved by updating
the 'next' index during a free to point to the location just freed.
This causes valloc to prefer to reuse regions before allocating
new ones, which saves physical memory. The one time this causes
extra work during an allocation is if something is freed at the start
of a long set of allocated blocks, which means the next allocation
would have to traverse the bitmap. But that's not too bad, as it's
probably rare, and once done once, will have the next pointer
updated to the end of that long list of allocated bits.