Fix signal handling for processes that catch them.
Signals that are caught set the THREAD_SIGNALED flag, which now
blocks other signals from being processed until the flag is cleared.
This prevents a race condition that could cause a thread to try to
handle a singal that was set to SIG_DFL.

Additionally, global_sig_mask inside the process structure is now
checked inside the check_signals function from sched.c, which now
correctly doesn't cancel the signal but instead just ignores it's
bit in the signals_pending queue until it gets unblocked.