Fix creating directories.
When a directory is created, the . and .. entries are created as
well. This is done with fs_link. fs_link checks to see if nlinks
is 1 when it's operating on a directory, and if so, prevents
creating new entries in it. This is for when a directory is partially
unlinked, but it broke directory creation. Now, fs_link takes an
additional bool argument allowing the caller to override that check.