Remove cpu->prev, and set THREAD_DEAD correctly.
Instead of waiting until after the thread has scheduled away and
relying on storing the 'previous' thread inside the CPU structure,
simply do all the work that required that right before the switch
occurs.

Previously, we relyed on this mechanism to raise the THREAD_DEAD
flag if the state was set to THREADSTATE_DEAD while we were
scheduling away from a thread. This way, exit could enqueue a thread
to be destroyed into the workqueue, and then test a final flag to make
sure that we don't free it while it's in use. This flag would be set
after the thread switched away, so it could never be run again.

Instead, we now just set the flag right before the switch. This is okay,
as long as the only times work is done from the workqueue is after the
scheduler has completed a schedule on that CPU, or inside the idle
thread for that CPU. No where else is permitted. This way, no work
can ever be done between the time the thread's flag gets raised, and
when is has finished switching to a different thread.