Fix blocking of threads, add blocking in mutex.
This is a significant change. So, be ready. We now use a seperate
structure for blocklists instead of just a linked list. This is
for abstraction, type checking, and ease of later modification.
Internally, it currently uses a linkedlist to store threads.

I added a wakeone function for blocklists which just wakes the head
of the list. There's no priority given to higher priority threads
at this time, but that would be a useful future feature.

There was a bug that was hard to find, which is why this work took
so long. In the linkedlist apply_head function, it would wait to
see the return value of fn, and then maybe delete the list element.
That was a problem, because when we unblocked in wakeone, we would
add the thread to the runqueue and set it to RUNNING, and THEN remove
it from the blocklist's linked list. If the thread were to block
again before it got removed, we would see the linked list try to
remove that element from a list that it wasn't a part of! Bad things!
So, now, the blocklists have their own lock, the linked list is
set to lockless, and the list element is removed before the thread
is restarted.

Finally, mutexes do actually block threads that wait on them. This
actually works now. Yay. I had to add a block_confirm function which
would set up the blocking for the task, then call a function to check
if it still wanted to block, and undo what it had done maybe. A
little messy, but it works.