Fix bufferlist corruption.
Turns out, if you're going to use a linked list to store buffers,
and use a different linked list each time you call the function,
and allow multiple callers to that function, you best use a linked
list that is allowed to allocate entries. Because if you don't,
you'll end up with multiple linked lists whose entries are sort-of
intermixed. Which is bad. Because that's a massive corruption,
filled with pain. Especially when they start deleting entries
that are in OTHER LISTS.

So, DON'T DO THAT.

YOU DUMMY.