Fix keyboard's stage2 handler.
The handler now reads in all of the keys stored in the buffer.
We have to do it atomically, since the stage1 handler could fire,
and we also have to assume that the stage2 handler is 'overcalled'.
As in, it may get called too many times per key. Because of this,
we have to conditionally subtract the index into the buffer.