Fix real hardware bugs.
First: the assertion in interrupt.c that checks for held_locks was
tripping because of the call to mm_mmap in main.c done by PID1
before it jumps to userspace. Make the assertion not trip if
THREAD_KERNEL is set (which is probably a good idea anyway), and
then set that flag in PID1 for the call to mmap (and unset it
afterwards).

Second: in pm_buddy, we need to make sure that we don't allow
freeing any memory that is above our "limit" from MAX_ORDER. Easy
little check to make sure.